Fix for about us fragment, now displaying on phone
