+++ app/src/main/java/org/codeforafrica/citizenreporterandroid/main/MainActivity.java
@@ -1,25 +1,20 @@
 package org.codeforafrica.citizenreporterandroid.main;
 
+import android.os.Bundle;
 import android.support.design.widget.TabLayout;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
-import android.os.Bundle;
-import android.view.LayoutInflater;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
-import android.view.ViewGroup;
-
-import android.widget.TextView;
+import android.widget.ScrollView;
 
 import org.codeforafrica.citizenreporterandroid.R;
+
+
 
 public class MainActivity extends AppCompatActivity {
 
@@ -100,6 +95,8 @@ public class MainActivity extends AppCompatActivity {
                     return AssignmentsFragment.newInstance();
                 case 1:
                     return StoriesFragment.newInstance();
+                case 2:
+                    return AboutFragment.newInstance();
                 default:
                     return null;
             }
@@ -107,8 +104,8 @@ public class MainActivity extends AppCompatActivity {
 
         @Override
         public int getCount() {
-            // Show 2 total pages.
-            return 2;
+            // Show 3 total pages.
+            return 3;
         }
 
         @Override
@@ -118,6 +115,8 @@ public class MainActivity extends AppCompatActivity {
                     return "Assignments";
                 case 1:
                     return "Stories";
+                case 2:
+                    return "About";
             }
             return null;
         }
